proyecto_3/Sistema_de_Gestión_de_Tareas/requirement.txt
coreapi==2.3.3
Django==5.0.1
djangorestframework==3.14.0